Add files via upload
This hands-on notebook provides a comprehensive understanding of how to work with PyTorch datasets and dataloaders, guiding us through their implementation from scratch. We begin by exploring the fundamental concepts of datasets and dataloaders, learning how they facilitate efficient data handling, batching, and processing in PyTorch. Through hands-on implementation, we gain practical experience in creating custom datasets and leveraging dataloaders to streamline data loading for model training. Once we establish a solid foundation in dataset and dataloader usage, we are introduced to tiktoken, a fast and efficient tokenizer designed to process text efficiently for deep learning applications. We then delve into the concept of positional token embeddings, which play a crucial role in encoding positional information within tokenized sequences. By embedding token IDs with positional information, we enable models to capture the order and structure of the input data effectively. This step-by-step approach ensures that we not only understand the theoretical aspects but also gain hands-on experience in implementing these key components. By the end of this notebook, we will have a thorough grasp of dataset management in PyTorch, efficient tokenization with tiktoken, and the importance of positional token embeddings in natural language processing tasks.